+++ hdsr_fewspy/api_calls/time_series/get_time_series_multi.py
@@ -82,7 +82,10 @@ class GetTimeSeriesMulti(GetTimeSeriesBase):
             all_file_paths.extend(file_paths_created)
             progress_percentage = int((index + 1) / nr_total * 100)
             logger.info(f"get_time_series_multi progress = {progress_percentage}%")
-        logger.info(f"finished download and writing to {len(all_file_paths)} file(s)")
+        if all_file_paths:
+            logger.info(f"finished download and writing to {len(all_file_paths)} file(s)")
+        else:
+            logger.warning("finished download but no data found, so nothing to write to file")
         return all_file_paths
 
     @classmethod