+++ HISTORY.rst
@@ -2,7 +2,7 @@
 ----------------
 - add logging to examples
 - add timeseries flag description (Deltares url) to readme
-- add __repr__ to PiSettings so that R users can print those settings
+- add __repr__ to PiSettings so that non-python users (R, etc.) can print those settings
 
 1.11 (2023-6-1)
 ----------------